requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
-e git+https://github.com/DH-IT-Portal-Development/django-shared-core.git@questions#egg=cdh-django-core[all]
    # via -r requirements.in
asgiref==3.7.2
    # via django
blessed==1.20.0
    # via curtsies
bpython==0.24
    # via
    #   -r requirements.in
    #   cdh-django-core
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via pip-tools
closure==20191111
    # via cdh-django-core
cryptography==41.0.4
    # via
    #   -r requirements.in
    #   cdh-django-core
    #   pyopenssl
    #   pysaml2
curtsies==0.4.2
    # via bpython
cwcwidth==0.1.9
    # via
    #   bpython
    #   curtsies
defusedxml==0.7.1
    # via
    #   djangosaml2
    #   pysaml2
deprecated==1.2.14
    # via cdh-django-core
django==4.2.6
    # via
    #   -r requirements.in
    #   cdh-django-core
    #   django-braces
    #   django-csp
    #   django-debug-toolbar
    #   django-extensions
    #   django-modeltranslation
    #   django-simple-menu
    #   djangorestframework
    #   djangosaml2
django-braces==1.15.0
    # via
    #   -r requirements.in
    #   cdh-django-core
django-csp==3.7
    # via
    #   -r requirements.in
    #   cdh-django-core
django-debug-toolbar==4.2.0
    # via
    #   -r requirements.in
    #   cdh-django-core
django-extensions==3.2.3
    # via
    #   -r requirements.in
    #   cdh-django-core
django-impersonate==1.9.1
    # via
    #   -r requirements.in
    #   cdh-django-core
django-modeltranslation==0.18.11
    # via
    #   -r requirements.in
    #   cdh-django-core
django-simple-menu==2.1.3
    # via
    #   -r requirements.in
    #   cdh-django-core
djangorestframework==3.14.0
    # via cdh-django-core
djangosaml2==1.7.0
    # via cdh-django-core
elementpath==4.1.5
    # via xmlschema
greenlet==3.0.0
    # via bpython
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via build
lesscpy==0.15.1
    # via cdh-django-core
mysqlclient==2.2.0
    # via -r requirements.in
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements.in
ply==3.11
    # via lesscpy
pscript==0.7.7
    # via vbuild
pycparser==2.21
    # via cffi
pygments==2.16.1
    # via bpython
pyjwt==2.8.0
    # via cdh-django-core
pyopenssl==23.2.0
    # via pysaml2
pyproject-hooks==1.0.0
    # via build
pysaml2==7.4.2
    # via djangosaml2
pyscss==1.4.0
    # via cdh-django-core
python-dateutil==2.8.2
    # via pysaml2
python-magic==0.4.27
    # via cdh-django-core
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   pysaml2
pyxdg==0.28
    # via bpython
requests==2.31.0
    # via
    #   bpython
    #   cdh-django-core
    #   pysaml2
six==1.16.0
    # via
    #   blessed
    #   pyscss
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.8.0
    # via
    #   asgiref
    #   django-modeltranslation
urllib3==2.0.6
    # via requests
vbuild==0.8.2
    # via cdh-django-core
wcwidth==0.2.8
    # via blessed
wheel==0.41.2
    # via pip-tools
wrapt==1.15.0
    # via deprecated
xmlschema==2.5.0
    # via pysaml2
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
